[tracking] Renaming command `tracking` to `config`
